Rename the source and updates pom to show its a component
