+++ tests/playwright/ui-tests/trusts/contacts-page.spec.ts
@@ -2,7 +2,7 @@ import { test } from '@playwright/test'
 import { ContactsPage } from '../../page-object-model/trust/contacts-page'
 import { NotFoundPage } from '../../page-object-model/not-found-page'
 import { FakeTestData } from '../../fake-data/fake-test-data'
-import { DataSourcePanelItem } from '../../page-object-model/trust/sources-and-updates'
+import { DataSourcePanelItem } from '../../page-object-model/trust/sources-and-updates-component'
 
 test.describe('Contacts page', () => {
   let contactsPage: ContactsPage
